lib/intel_chipset: Add Lunar Lake support
Add basic definition for the Lunar Lake platform. This uses the PCI ID
from the xe_pciids.h which has a macro in a slighlty different form.

Going forward we could also trim down a little bit the the info struct
like was done in the kernel, relying on gmdid and IP version checks.
This is left for later.

v2: Remove has_gmdid that is still not used

Cc: Swati Sharma <swati2.sharma@intel.com>
Signed-off-by: Lucas De Marchi <lucas.demarchi@intel.com>
Reviewed-by: Balasubramani Vivekanandan <balasubramani.vivekanandan@intel.com>