igt_core: Show the end of each subtest in dmesg
We have an igt_kmsg for the start of each test, and so we know
the end by looking for the next start message or the end of the
whole test. However, logs are split into each subtest and it is
not clear in the kernel message where or even _if_ the test
finished executing.

Signed-off-by: Chris Wilson <chris.p.wilson@linux.intel.com>
Signed-off-by: Kamil Konieczny <kamil.konieczny@linux.intel.com>
Reviewed-by: Zbigniew Kempczyński <zbigniew.kempczynski@intel.com>