igt_core: Avoid exit_subtest in subthread
If a thread crashes, the signal handler for SIGSEGV and pals try to
make the subtest fail and eventually longjmps out of the
subtest. Naturally this can cause multiple threads to yield subtest
results and other weirdness.

We already check for non-main-thread in igt_fail_assert() but the path
from a signal handler to subtest exit doesn't have such protection.

Cc: Petri Latvala <adrinael@adrinael.net>
Signed-off-by: Petri Latvala <petri.latvala@intel.com>
Signed-off-by: Kamil Konieczny <kamil.konieczny@linux.intel.com>
Reviewed-by: Zbigniew Kempczyński <zbigniew.kempczynski@intel.com>