tests/xe: remove an useless TODO message
When I originally wrote the test descriptions, I tought it
was a good idea to add a warn message to warn about
run type == FULL.

The documentation logic and tools improved a lot since them and
listing those are really easy nowadays. So, remove the
TODOs with:

	TODO: change ``'Run type' == FULL`` to a better category

in order to cleanup the documentation.

Reviewed-by: Kamil Konieczny <kamil.konieczny@linux.intel.com>
Signed-off-by: Mauro Carvalho Chehab <mchehab@kernel.org>