scripts/test_list.py: speedup testlist check logic
Now that the build system will generate the testlist for us,
just read them when checking for missing symbols.

This should speed up the check logic.

After the change, touching on a single file and rebuilding the
testplan with check enabled faster. On a machine with SSD disks:

Before (flushing disk caches):

	$ sync; sudo su -c "echo 3 > /proc/sys/vm/drop_caches"; rm -rf build/tests/i915_hangman*; time ninja -C build
	ninja: Entering directory `build'
	[10/10] Generating docs/testplan/i915_tests.html with a custom command

	real	0m7.669s
	user	0m8.050s
	sys	0m1.360s

After:
	$ sync; sudo su -c "echo 3 > /proc/sys/vm/drop_caches"; rm -rf build/tests/i915_hangman*; time ninja -C build
	ninja: Entering directory `build'
	[10/10] Generating docs/testplan/i915_tests.html with a custom command

	real	0m5.329s
	user	0m6.415s
	sys	0m0.271s

In practice, results should be a lot faster depending on disk cache,
and disk speed.

This can be improved even further if we split the test_executables
dependencies per driver.

Reviewed-by: Kamil Konieczny <kamil.konieczny@linux.intel.com>
Signed-off-by: Mauro Carvalho Chehab <mchehab@kernel.org>