test/i915/gem_create: Skip the test if the PAT index is set on a non MTL device
If we try to set the PAT index on a device that is not a Meteor
Lake, we expect an -ENODEV return value. But then, we forget to
skip and soon after we assert on the return value being '0',
forcing the error path.

Skip, as expected, in case the return value is either EINVAL or
ENODEV, the latter for meteor lake devices.

Suggested-by: Tvrtko Ursulin <tvrtko.ursulin@linux.intel.com>
Signed-off-by: Andi Shyti <andi.shyti@linux.intel.com>
Cc: Fei Yang <fei.yang@intel.com>
Cc: Kamil Konieczny <kamil.konieczny@linux.intel.com>
Reviewed-by: Tvrtko Ursulin <tvrtko.ursulin@intel.com>
Reviewed-by: Fei Yang <fei.yang@intel.com>