tests/gem_ccs: Drop local function for setting ctrl surface object
Until now gem_ccs was the only user of set_surf_object() and probably
from this reason it wasn't "globalized". Upcoming gem_ccs counterpart
for xe - xe_ccs - uses same object initialization and we got this
function (with small rename) in the blitter library so we can use it
and get rid of local version.

Signed-off-by: Zbigniew Kempczyński <zbigniew.kempczynski@intel.com>
Reviewed-by: Karolina Stolarek <karolina.stolarek@intel.com>