xe: Rename engine to exec_queue
Engine was inappropriately used to refer to execution queues and
it also created some confusion with hardware engines. Update to
the kernel header where this was fixed and also update variable
names in tests.

The xe_drm.h uapi header is now in sync with drm-xe-next
commit ("drm/xe: Rename engine to exec_queue").

Signed-off-by: Francois Dugast <francois.dugast@intel.com>
Reviewed-by: Matthew Brost <matthew.brost@intel.com>
Signed-off-by: Rodrigo Vivi <rodrigo.vivi@intel.com>