tests/amdgpu: Add test for MALL static screen cache scanout
MALL stands for memory access at last level.

This test is to verify that the display scanout is triggered from MALL cache instead
of GPU VRAM when the contents of the screen is idle for some time.

Requires UMR userspace utility from
https://gitlab.freedesktop.org/tomstdenis/umr to read MALL register
value.

This test commits a test pattern FB, sleeps for some
time, and then reads back the MALL status register to check if display
refresh from static screen is triggered.

Changes in v2:

Added possibility to skip the test on Asics that do not support MALL

Changes in v3:

Aded comment in source code about the MALL delay time necessary to
trigger idle optimizations.

Signed-off-by: Aurabindo Pillai <aurabindo.pillai@amd.com>
Reviewed-by: Alex Hung <alex.hung@amd.com>