tests/kms_plane: Skip plane position test if primary ID > sprite ID
Currently, IGT assumes that the primary plane in
test_plane_position_with_output() will always be committed under the
sprite plane. This will not always hold true if the plane_id of the
primary plane is greater than the plane_id of the sprite as the DRM
framework will calculate a higher normalized-zpos for planes with larger
plane_id's if the zpos property isn't set [1].

Since this is a non-atomic test, we won't be able to explicitly set the
zpos property. So instead, skip testing the plane position test output for
cases where the primary plane ID is greater than the sprite plane ID

[1] https://elixir.bootlin.com/linux/v6.5-rc5/source/drivers/gpu/drm/drm_blend.c#L431

Signed-off-by: Jessica Zhang <quic_jesszhan@quicinc.com>
Reviewed-by: Abhinav Kumar <quic_abhinavk@quicinc.com>