tests/kms_writeback: allow writeback to output custom format
In some display pipelines, such as MSM, the writeback block is
capable of writing in an output format other than DRM_FORMAT_XRGB8888
such as DRM_FORMAT_NV12.

To allow validation of these formats, allow kms_writeback to
take a command line argument which supports passing the writeback's
output fb format to something other than DRM_FORMAT_XRGB8888.

The DRM_FORMAT_XRGB8888 checks have been relaxed to allow YUV formats
as those are the ones verified while adding support for this feature
with MSM boards.

This option is supported only along with the --dump option because
the CRC validation API needs to be extended to handle
non-DRM_FORMAT_XRGB8888 formats to make that possible.

changes in v2:
	- removed an unnecessary new line

changes in v3:
	- renamed wb_yuv_fmt to wb_fmt to allow all formats
	- restrict the custom format usage to --dump option

Signed-off-by: Abhinav Kumar <quic_abhinavk@quicinc.com>
[Added Petri's R-b as agreed on IRC]
Reviewed-by: Petri Latvala <adrinael@adrinael.net>