tests/xe: Add a test that validates residency during s2idle
During s2idle, the device is powered up and GT C6 enabled
so we expect the residency to increase across suspend.
Add a test that validates the residency over a s2idle cycle
and checks if it is greater than the autoresume delay
or within tolerance.

v2: skip d3cold disable for igfx (Anshuman)
v3: rename test flag to TEST_IDLE
    reorder variables (Anshuman)
v4: check if residency is greater than autoresume delay as
    expected. Assert only if residency is lesser than autoresume
    delay and not within tolerance.

Signed-off-by: Riana Tauro <riana.tauro@intel.com>
Reviewed-by: Anshuman Gupta <anshuman.gupta@intel.com>
Signed-off-by: Anshuman Gupta <anshuman.gupta@intel.com>