tests/kms_atomic: remove zpos < N-planes assert
Generalize plane-immutable-zpos test case by removing not-required
assert. zpos < N-planes is a condition not required by DRM
documentation. Removing this validation doesn't affect the test purpose
on verifying that planes with higher zpos always cover those with lower
zpos. Also, this assert doesn't fit driver's configuration where the
range of plane zpos is just defined by the plane type but its max value
doesn't need to be normalized to the number of planes in a CRTC, for
example, a cursor plane with the highest possible zpos value.

Signed-off-by: Melissa Wen <mwen@igalia.com>
Reviewed-by: Rodrigo Siqueira <Rodrigo.Siqueira@amd.com>