tests/kms_atomic_transition: Address test flakiness
Nonblocking commits can sometimes return -EBUSY. It's expected that
this error is transient and that userspace should retry the commit.

Since kms_atomic_transition uses nonblocking commits for some test
cases, the test fails flakily.

This change will handle -EBUSY to ensure that the test does not
fail in that case.

Closes: https://gitlab.freedesktop.org/drm/igt-gpu-tools/-/issues/144
Signed-off-by: Drew Davenport <ddavenport@chromium.org>
Reviewed-by: Juha-Pekka Heikkila <juhapekka.heikkila@gmail.com>