lib/intel_compute: Add name field for debugging purposes
Debugging without knowledge about object characteristics is hard and
time consuming. Simple name field added for printing binded addresses
and their sizes might speed up development. I experienced this on
extending to DG2 so I decided to permanently add it. But to avoid
annoying output this is limited to igt_debug() which will print
only on user request or on the test failure.

Signed-off-by: Zbigniew Kempczyński <zbigniew.kempczynski@intel.com>
Cc: Christoph Manszewski <christoph.manszewski@intel.com>
Cc: Francois Dugast <francois.dugast@intel.com>
Cc: Mauro Carvalho Chehab <mchehab@kernel.org>
Reviewed-by: Francois Dugast <francois.dugast@intel.com>