tests/xe: Add a test that validates idle residency on exec
Add a test what runs a background load that is
active approximately 1% of the time. Verify that we do enter
GT-C6 the rest of the time and validate idle residency is within
tolerance.

v2: use start and end variables for residency
    add detailed comment (Badal)

v3: run on a single instance of an engine belonging to gt (Anshuman)

Signed-off-by: Riana Tauro <riana.tauro@intel.com>
Reviewed-by: Badal Nilawar <badal.nilawar@intel.com>