lib/intel_mocs: Stop encoding mocs
From gen9 until xe2, the various gpu commands use a mocs object that is
defined as

	[0]   pxp
	[6:1] mocs index

However this is not true for xe2 anymore. Also the way it is coded right
now it's very easy to double shift the mocs value and end up with the
wrong setting. Avoid this by separating where the index should be used
(intel_mocs.[ch]) and the encoding on the various commands. To help with
clarity the following renames were done:

	intel_get_wb_mocs() -> intel_get_wb_mocs_index()
	intel_get_uc_mocs() -> intel_get_uc_mocs_index()
	gen9_surface_state.*.memory_object_control
		 -> gen9_surface_state.*.mocs_index
	mocs -> mocs_index in lib/intel_blt.*

v2:
  - Use I915_MOCS_PTE instead of the hardcoded 1 in the gen11 media
    function (Matt Roper)
  - Fix a wrong shift by XY_FAST_COLOR_BLT_MOCS_INDEX_SHIFT, result of a
    copy and paste mistake from lib/igt_draw.c to lib/igt_fb.c
    (Matt Roper)
  - In addition to the conversions in intel_blt, also add a _index
    suffix to avoid confusion

Signed-off-by: Lucas De Marchi <lucas.demarchi@intel.com>
Reviewed-by: Matt Roper <matthew.d.roper@intel.com>