tests/amd_dispatch: add negative test
Dispatch a bad program on gfx/compute ring.
Check whether they will hang.
Expose the additional parameter 'hung' to the upper layer.
Dispatch existent a binary shader that waits for the register to be changed,
but it does not occur and as a result, we have GPU reset and check this.

V3:
  -add detail description and fix code style(Kamil)
  -improve test description(Vitaly)

Cc: Vitaly Prosyak <vitaly.prosyak@amd.com>
Cc: Luben Tuikov <luben.tuikov@amd.com>
Cc: Alex Deucher <alexander.deucher@amd.com>
Cc: Christian Koenig <christian.koenig@amd.com>
Cc: Kamil Konieczny <kamil.konieczny@linux.intel.com>
Signed-off-by: Jesse zhang <jesse.zhang@amd.com>
Reviewed-by: Vitaly Prosyak <vitaly.prosyak@amd.com>