scripts/test_list.py: better name testlist include/exclude config fields
The config names "testlists" and "blocklists" are confusing, as they
don't have the same concept of such lists used within IGT.

What the test_list logic does, instead, is to fill a field with
a value for all tests included from a file, except for the ones
marked on an exclude testlist.

Use the name "include" for the tag defining the file name
containing a list of tests to be included.

On a similar way, use "exclude" to mean the file name containg
a list of tests to be excluded.

While here, remove two class variables originally planned for
the testlist/blocklist logic that aren't actually used.

Reviewed-by: Kamil Konieczny <kamil.konieczny@linux.intel.com>
Signed-off-by: Mauro Carvalho Chehab <mchehab@kernel.org>