tests/kms_plane_scaling: use get_num_scalers()
Verify the count of supported scalers per CRTC, and proceed with
executing tests if the necessary number of scalers is available.
With the exception of "planes-.*" and "2x-scaler-multi-pipe"
subtests, a minimum of 1 scaler is needed for all other tests,
while these specific tests require a minimum of 2 scalers.

v2: -improved commit message (Kamil)
    -handled get_num_scalers() for non-intel devices (Kamil)
v3: -fixed get_num_scalers() (JP)
    -added get_num_scalers() to lib (JP)

Signed-off-by: Swati Sharma <swati2.sharma@intel.com>