tests/device_reset: fix logic to unbind when snd_hda_intel is loaded
For i915 devices are used, there's a glue with snd_hda_intel.
Currently, such glue requires that the audio driver to be unused
and removed before being able to remove the i915 driver.

There is already a logic at igt library to do it the right way,
but device_reset currently doesn't use.

Also, the logic is hardcoded to work only with Haswell, Broadwell
and DG1.

Change the logic to use the proper logic to remove such
module using the library.

Reviewed-by: Andrzej Hajda <andrzej.hajda@intel.com>
Reviewed-by: Nirmoy Das <nirmoy.das@intel.com>
Signed-off-by: Mauro Carvalho Chehab <mchehab@kernel.org>