lib/igt_draw: Check for Rendercopy support
igt_draw_supports_method() is initially designed to identify
that the given driver (i915/xe) supports the selected draw
method.

Even though Rendercopy support in IGT is not present for few
latest platforms, this api assumes that Rendercopy is present.
Add a check if Rendercopy supports on given platform or not.

Cc: Juha-Pekka Heikkila <juhapekka.heikkila@gmail.com>
Signed-off-by: Bhanuprakash Modem <bhanuprakash.modem@intel.com>
Reviewed-by: Juha-Pekka Heikkila <juhapekka.heikkila@gmail.com>
Reviewed-by: Kamil Konieczny <kamil.konieczny@linux.intel.com>