scripts/igt_doc.py: add a logic to generate Intel CI testlists
Testlist for Intel CI requires parsing not only IGT testlists,
but also block and permit lists. This is currently handled
internally, but not from the documentation.

Now that we have everything set in place, add a method for
generating it at IGT. The logic there is somewhat generic,
but it expects some fields with a namespace as defined on
tests/intel/*.json files.

So, instead of placing at the generic code (test_list.py),
add them to igt_doc.py, where IGT-specific glue can easily
be added while keeping test_list.py generic enought to be
used on other projects.

Signed-off-by: Mauro Carvalho Chehab <mchehab@kernel.org>
Acked-by: Kamil Konieczny <kamil.konieczny@linux.intel.com>