tests/kms_writeback: fix kms_writeback failure for XRGB8888 only devices
commit ac56ba97248b ("tests/kms_writeback: support DRM_FORMAT_XRGB2101010
for writeback") seems to have broken kms_writeback for devices which
support only XRGB8888 because check_writeback_config() will return a value
based on the last format in the fourcc[] array.

For devices such as MSM which do not support DRM_FORMAT_XRGB2101010, this
will return an error and overall kms_writeback_get_output() will not
return a valid output for the writeback connector.

Change the return of check_writeback_config() to return based on the
supported_colors bitmask so that only when neither XRGB8888 or
DRM_FORMAT_XRGB2101010 is supported, then the test is skipped.

Signed-off-by: Abhinav Kumar <quic_abhinavk@quicinc.com>
Reviewed-by: Alex Hung <alex.hung@amd.com>